Change template_general so that IE loads in standards mode.  Fixes #18.
The HTML comment with the documentation for developers was tricking IE into loading with Quirks mode.
Moving the documentation (temporarily to READEME.md) and having the first line in the rendered HTML page be the <!doctype html> declaration puts IE back into Standards Mode.
Consider including documentation as a Twig comment?