Compact stylesheets. Ref #6.
Minifying and including flexslider.css and style.css into the main
style-min-XXXXXX.css causes list the list icons in the Flexslider to
appear vertically instead of where the belong along the bottom of the
slideshow.
As such, only roll in normalize.css, grid.css, and superfish.css into
the main minified css file.