Fix player info field translation issue.
Before the fix of the great Lagwagon of 2014, both the Guest and Player spreadsheets reported their rows with the same field names.  In that instance, a single dictionary for translating Google fieldnames to HULA fieldnames was sufficient.
After the fix of the great Lagwagon, the player sheet still reports with the old fieldnames, but the Guest spreadsheet now returns random-character fieldnames (all starting with an underscore).
This commit maps those underscore fieldnames to HULA fieldnames and adds a check for which dictionary to use.

Party on Wayne.