Fix error resulting from comma in team name.
Request was being made on the string 'SS_KEY,Teamname' by splitting on the comma.  Then the request was built on (array[0], array[1]), but a Teamname with a comma in it would be truncated.  Now use slice to capture all parts of the string after the SS_KEY is chopped off.