Revert "Fix player info field translation issue."
This reverts commit 96f7f7bc991d69ae003404d53c8c76b91c4d1c34.

The bug that this commit was supposed to fix has become intermittent.  Now the response from Google Sheets only seems to come with underscore-prefixed field names under the following condition:
+ Load page
+ Select Guest

But this only sometimes reproduces the bug.

Other times, and with any other workflow tried, Google sheet responds with the same fieldnames as with the Player sheets.  Thus the multiple dictionary approach is being scrapped until the problem can actually be figured out.